Refactor: change pageSlice to paramSlice
