+++ src/pages/AccountList/components/PageButtons.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
 import { useSelector } from 'react-redux';
-import { updatePage } from '../../../store/services/pageSlice';
+import { updateParams } from '../../../store/services/paramSlice';
 import { ReducerType } from '../../../store/rootReducer';
 import { getAccountList } from '../../../store/services/accountSlice';
 import { AppDispatch } from '../../../store/';
 
 const PageButton = () => {
-  const { params, accountPages } = useSelector((state: ReducerType) => state.page);
+  const { params, accountPages } = useSelector((state: ReducerType) => state.param);
   const dispatch = useDispatch<AppDispatch>();
 
   const changePage: React.MouseEventHandler<HTMLButtonElement> = e => {
     const { value } = e.target as HTMLButtonElement;
-    dispatch(updatePage(value));
+    dispatch(updateParams(value));
     dispatch(getAccountList(params));
   };
